OpenJudge/pyproject.toml
[project]
name = "coughoverflow-yuechenghao626"
version = "0.1.0"
description = ""
authors = [
    {name = "Yuecheng Hao",email = "hyc626123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "celery[sqs] (>=5.5.2,<6.0.0)",
    "boto3 (>=1.38.13,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pyjwt (>=2.8.0,<3.0.0)",
    "bcrypt (>=4.0.0,<5.0.0)",
    "wtforms (>=3.0.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
